# app/Http/Controllers/Admin/CategoryController.php
<?php

namespace App\Http\Controllers\Admin;

use App\Http\Controllers\Controller;
use App\Http\Requests\Category\StoreCategoryRequest;
use App\Http\Requests\Category\UpdateCategoryRequest;
use App\Http\Resources\CategoryResource;
use App\Models\Category;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
use Illuminate\Support\Facades\Storage;
use OpenApi\Attributes as OA;

class CategoryController extends Controller
{
    #[OA\Get(
        path: "/api/admin/categories",
        summary: "Get list of categories with filters",
        tags: ["Admin - Categories"],
        security: [["apiAuth" => []]],
        parameters: [
            new OA\Parameter(name: "name", in: "query", schema: new OA\Schema(type: "string")),
            new OA\Parameter(name: "main_category_id", in: "query", schema: new OA\Schema(type: "integer")),
            new OA\Parameter(name: "category_type", in: "query", schema: new OA\Schema(type: "string")),
            new OA\Parameter(name: "featured_order", in: "query", schema: new OA\Schema(type: "integer")),
            new OA\Parameter(name: "limit", in: "query", schema: new OA\Schema(type: "integer", default: 15)),
            new OA\Parameter(name: "offset", in: "query", schema: new OA\Schema(type: "integer", default: 0)),
        ],
        responses: [
            new OA\Response(response: 200, description: "List of categories", content: new OA\JsonContent(type: "array", items: new OA\Items(ref: "#/components/schemas/CategoryResource")))
        ]
    )]
    public function index(Request $request): AnonymousResourceCollection
    {
        $query = Category::with(['mainCategory']);

        if ($request->filled('name')) {
            $query->where('name', 'like', '%' . $request->name . '%');
        }

        if ($request->filled('main_category_id')) {
            $query->where('main_category_id', $request->main_category_id);
        }

        if ($request->filled('category_type')) {
            $query->where('category_type', $request->category_type);
        }

        if ($request->filled('featured_order')) {
            $query->where('featured_order', $request->featured_order);
        }

        $categories = $query->orderBy('featured_order', 'asc')
            ->orderBy('id', 'desc')
            ->skip($request->integer('offset', 0))
            ->take($request->integer('limit', 15))
            ->get();

        return CategoryResource::collection($categories);
    }

    #[OA\Post(
        path: "/api/admin/categories",
        summary: "Create a new category",
        tags: ["Admin - Categories"],
        security: [["apiAuth" => []]],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\MediaType(
                mediaType: "multipart/form-data",
                schema: new OA\Schema(
                    required: ["name", "main_category_id"],
                    properties: [
                        new OA\Property(property: "name", type: "string"),
                        new OA\Property(property: "main_category_id", type: "integer"),
                        new OA\Property(property: "category_type", type: "string"),
                        new OA\Property(property: "slug", type: "string"),
                        new OA\Property(property: "is_active", type: "boolean"),
                        new OA\Property(property: "featured_order", type: "integer"),
                    ]
                )
            )
        ),
        responses: [
            new OA\Response(response: 201, description: "Category created", content: new OA\JsonContent(ref: "#/components/schemas/CategoryResource")),
            new OA\Response(response: 422, description: "Validation error")
        ]
    )]
    public function store(StoreCategoryRequest $request): JsonResponse
    {
        $data = $request->validated();

        $category = Category::create($data);

        return (new CategoryResource($category->load('mainCategory')))
            ->response()
            ->setStatusCode(201);
    }

    #[OA\Get(
        path: "/api/admin/categories/{id}",
        summary: "Get single category",
        tags: ["Admin - Categories"],
        security: [["apiAuth" => []]],
        responses: [
            new OA\Response(response: 200, description: "Category details", content: new OA\JsonContent(ref: "#/components/schemas/CategoryResource")),
            new OA\Response(response: 404, description: "Category not found")
        ]
    )]
    public function show(Category $category): CategoryResource
    {
        return new CategoryResource($category->load(['mainCategory', 'subCategories']));
    }

    #[OA\Put(
        path: "/api/admin/categories/{id}",
        summary: "Update category",
        description: "Use POST with _method=PUT for multipart/form-data updates",
        tags: ["Admin - Categories"],
        security: [["apiAuth" => []]],
        parameters: [new OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "integer"))],
        requestBody: new OA\RequestBody(
            content: new OA\MediaType(
                mediaType: "multipart/form-data",
                schema: new OA\Schema(
                    properties: [
                        new OA\Property(property: "_method", type: "string", example: "PUT"),
                        new OA\Property(property: "name", type: "string"),
                        new OA\Property(property: "main_category_id", type: "integer"),
                        new OA\Property(property: "category_type", type: "string"),
                        new OA\Property(property: "slug", type: "string"),
                        new OA\Property(property: "is_active", type: "boolean"),
                        new OA\Property(property: "featured_order", type: "integer"),
                    ]
                )
            )
        ),
        responses: [
            new OA\Response(response: 200, description: "Category updated", content: new OA\JsonContent(ref: "#/components/schemas/CategoryResource")),
            new OA\Response(response: 404, description: "Category not found")
        ]
    )]
    public function update(UpdateCategoryRequest $request, Category $category): CategoryResource
    {
        $data = $request->validated();

        $category->update($data);

        return new CategoryResource($category);
    }

    #[OA\Delete(
        path: "/api/admin/categories/{id}",
        summary: "Delete category",
        tags: ["Admin - Categories"],
        security: [["apiAuth" => []]],
        responses: [
            new OA\Response(response: 204, description: "Category deleted")
        ]
    )]
    public function destroy(Category $category): JsonResponse
    {
        $category->delete();

        return response()->json(null, 204);
    }

    #[OA\Post(
        path: "/api/admin/categories/reorder",
        summary: "Batch update categories order/blocks",
        tags: ["Admin - Categories"],
        security: [["apiAuth" => []]],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                properties: [
                    new OA\Property(property: "items", type: "array", items: new OA\Items(
                        properties: [
                            new OA\Property(property: "id", type: "integer"),
                            new OA\Property(property: "featured_order", type: "integer"),
                        ]
                    ))
                ]
            )
        ),
        responses: [
            new OA\Response(response: 200, description: "Order updated")
        ]
    )]
    public function reorder(Request $request): JsonResponse
    {
        $request->validate([
            'items' => ['required', 'array'],
            'items.*.id' => ['required', 'exists:categories,id'],
            'items.*.featured_order' => ['nullable', 'integer'],
        ]);

        foreach ($request->items as $item) {
            $category = Category::find($item['id']);
            $category->update(array_filter($item, fn($key) => $key !== 'id', ARRAY_FILTER_USE_KEY));
        }

        return response()->json(['message' => 'Categories updated successfully']);
    }
}